backend/school_system/serializers.py

from rest_framework import serializers
from authentication.models import School


class SchoolSerializer(serializers.ModelSerializer):
    class Meta:
        model = School
        fields = ['school_id', 'school_name']
