backend/backend_requirements.txt

Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.5
mysql-connector-python==9.1.0
python-dotenv==1.0.1
bcrypt==4.2.1
requests==2.32.3
pandas==2.2.3
